+++ Runtime/Scripts/IngameDebug/Console/CommandDisplayOptions.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace ANU.IngameDebug.Console
+{
+    [Flags]
+    public enum CommandDisplayOptions
+    {
+        Console = 1 << 0,
+        Dashboard = 1 << 1,
+    }
+}