+++ Runtime/Scripts/IngameDebug/Console/CommandRegistrationType.cs
@@ -1,0 +1,8 @@
+namespace ANU.IngameDebug.Console
+{
+    public enum CommandRegistrationType
+    {
+        Synchronous,
+        Asynchronous
+    }
+}